Final organization, bugfixing (printing probabilities at every step)
